pyproject.toml
[project]
name = "prediction-survival"
version = "1.1.0"
description = "Survivor prediction using classical machine learning models and a deep neural network."
readme = "README.md"
requires-python = ">=3.11,<3.12"
authors = [{name = "Jeremy Vachier"}]
dependencies = [
	"numpy>=1.26.2",
	"pandas>=2.1.4",
	"scikit-learn>=1.5.0",
	"tensorflow>=2.15.0; sys_platform != 'darwin'",
	"tensorflow-macos>=2.15.0; sys_platform == 'darwin'",
	"tensorflow-io-gcs-filesystem>=0.34.0; sys_platform == 'darwin'",
	"xgboost>=2.0.0",
	"lightgbm>=4.0.0",
	"catboost>=1.2.0",
	"pyyaml>=6.0",
	"joblib>=1.3.0",
]

[dependency-groups]
dev = [
	"black>=24.3.0",
	"ruff>=0.2.1",
	"pylint>=3.0.3",
	"pytest>=7.4.0",
	"pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
